requirements.txt
# Core dependencies
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.25.0

# E-ink display library
omni-epd>=0.4.0

# Web scraping and rendering
selenium>=4.0.0

# Optional: for better performance on Raspberry Pi
# RPi.GPIO>=0.7.0  # Will be installed by omni-epd if on Pi
# spidev>=3.5      # Will be installed by omni-epd if on Pi

# Development and testing
pytest>=6.0.0
